+++ src/main/java/com/demcha/compose/document/chart/ChartStyle.java
@@ -1,6 +1,7 @@
 package com.demcha.compose.document.chart;
 
 import com.demcha.compose.document.style.DocumentCornerRadius;
+import com.demcha.compose.document.style.DocumentPaint;
 import com.demcha.compose.document.style.DocumentStroke;
 import com.demcha.compose.document.style.DocumentTextStyle;